pyproject.toml
[project]
name = "phote-meta-organizer"
version = "0.2.0"
description = "照片元数据整理工具 - 按时间整理、修复元数据、批量重命名"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "piexif>=1.1.3",
    "pillow>=12.0.0",
    "pyyaml>=6.0",
    "pillow-heif>=0.20.0",
    "typer>=0.20.1",
]

[project.scripts]
photo-organizer = "photo_meta_organizer.cli:main"
pmo = "photo_meta_organizer.cli:main"

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
